regulatory-kb/requirements.txt
mkdocs-material>=9.5
pyyaml>=6.0
jsonschema>=4.0
python-frontmatter>=1.0
